Update versions and divide CallWrapper into interface and factory.
