+++ retromock/src/main/java/co/infinum/retromock/CallWrapper.java
@@ -1,0 +1,10 @@
+package co.infinum.retromock;
+
+import java.lang.reflect.Type;
+
+interface CallWrapper {
+
+  Object wrap(final Object call);
+
+  Type getReturnType();
+}
