+++ retromock/versions.gradle
@@ -6,8 +6,8 @@ buildscript {
             "junit"    : "5.1.1",
             "assertj"  : "3.8.0",
             "mockito"  : "2.23.0",
-            "kotlin"   : "1.3.61",
-            "coroutines" : "1.3.3",
+            "kotlin"   : "1.3.72",
+            "coroutines" : "1.3.6",
             "google"   : [
                     "annotations": "3.0.1"
             ]
